CameraHelpers: handle jumpTo
